backend/requirements.txt
fastapi
uvicorn
openai
pydantic-settings
requests